* Refactor apis to routes and controller
