+++ controllers/usersContoller.js
@@ -1,0 +1,36 @@
+const addUser = (req, res) => {
+  return res.status(500).json({
+    status: 'fail',
+    message: 'Not implemented'
+  })
+}
+
+const getUser = (req, res) => {
+  return res.status(500).json({
+    status: 'fail',
+    message: 'Not implemented'
+  })
+}
+
+const getUsers = (req, res) => {
+  return res.status(500).json({
+    status: 'fail',
+    message: 'Not implemented'
+  })
+}
+
+const updateUser = (req, res) => {
+  return res.status(500).json({
+    status: 'fail',
+    message: 'Not implemented'
+  })
+}
+
+const deleteUser = (req, res) => {
+  return res.status(500).json({
+    status: 'fail',
+    message: 'Not implemented'
+  })
+}
+
+module.exports = { getUser, getUsers, addUser, updateUser, deleteUser }
